trace: add sampling
If no configuration is specified, a Tracer defaults
to sampling all transactions. Otherwise, if the
ELASTIC_APM_TRANSACTION_SAMPLE_RATE environment is
specified, it is expected to be a ratio in the range
[0,1.0], and a sampler will be constructed that
samples pseudo-randomly (with a uniform distribution)
based on that ratio.

It is possible to change the sampler after the tracer
has been constructed using its SetSampler method.