contrib/apmlambda: intercept/trace AWS lambda
Importing this package will run our own net/rpc
listener/server, and override the environment
variable such that lambda.Start will not attempt
to listen on the same port. We then proxy the
requests through to the function registered with
lambda.Start, tracing them on the way.

Still TODO: provide an apmlambda.Start function
which can be used to add the transaction to the
context provided by lambda.