contrib/apmgin: don't create spans; set framework
Don't create a span when tracing the transaction.
It's not necessary, and not expected; only operations
within the HTTP request should result in spans being
created.

Also, allow the Tracer in Middleware to be nil,
meaning we'll use elasticapm.DefaultTracer. Also,
if DefaultTracer.Service.Framework is nil, set it
to apmgin.Framework during init().