transport: ELASTIC_APM_DEBUG=tracetransport=1
Introduce ELASTIC_APM_DEBUG, which can be used to
debug the APM code. The value can be a comma-separated
list of name=value pairs. Initially, the only name
that is understood is "tracetransport"; the value
is currently ignored, but users should supply "1"
for forwards compatibility.

Setting tracetransport=1 will cause the default
transport to be wrapped so that SendTransactions
and SendErrors are logged to stderr.