elasticapm: add MetricsGatherer
MetricsGatherer is an interface for
gathering metrics. You add one to a
Tracer using AddMetricsGatherer, and
it is periodically called to gather
all the metrics. Once they are
gathered, they are sent in a payload
to the Elastic APM server.

Mostly for testing purposes, you
can force metrics to be gathered and
sent at any given time by calling
Tracer.SendMetrics. It has the same
sort of behaviour as Tracer.Flush,
blocking until the metrics are sent
or the abort channel is signaled.