elasticapm: introduce Context type
This type stores transaction/error context, providing
methods for populating from HTTP request and response
data. Context objects are attached to Transactions or
Errors, and will be pooled along with them. Furthermore,
space for custom context will be reused, and we will
extend this to tags and other fields in the future.