elasticapm: validate/sanitize service name
When specifying the service name via environment
variable, it is now sanitized; invalid characters
are replaced with '_'. When creating a tracer
with NewTracer with an explicitly specified
service name, that service name must be valid,
or an error will be returned. Otherwise, the
sanitized executable name will be used as the
service name.