all: drop support for Service.Framework
Stop exposing the full service model in the
tracer - expose only the service name, version,
and environment instead. It is now no longer
possible to override the agent, language or
runtime; and it is no longer possible to set
the framework at all.

Framework is problematic because it doesn't
necessarily apply to all transactions, and
there's no way to attach the information at
the transaction level. So instead of including
it in a problematic state, we'll drop support
initially and work updates into the API and
server, and then reintroduce it later.

See https://github.com/elastic/apm-agent-go/issues/69