Support ELASTIC_APM_ACTIVE
If the tracer is inactive, then no transactons or errors will
be sent to the server; the tracer goroutine is never started.
The tracer grows an "Active() bool" method, which modules can
use to check if they should bail early.

Also, add a "RequestIgnorer" option to apmhttp, so clients can
programatically ignore certain requests.