module/...: introduce apmhttp.RequestWithContext
net/http.Request.WithContext has an unfortunate code
path which copies the *net.URL within the request,
which causes an allocation. We don't need (or want)
this behaviour, we just want to set the new context.
To avoid the allocation we introduce a new function
apmhttp.RequestWithContext, which avoids the copy by
clearing the request's URL during the WithContext
call.

Also, apmhttp.SetTransactionContext has been updated
slightly to drop the ResponseWriter parameter. The
Response type grows a Headers field, which is used
instead.