elasticapm: Done->End, Span.Start->Timestamp
Rename Span.Done and Transaction.Done methods to End.
Drop the time.Duration parameter, and instead expose
a time.Duration field on the types, initialised to
a negative value. If it is negative at the time End
is called, set it to the time elapsed since the
Timestamp field.

Also, rename Span.Start to Span.Timestamp to be
consistent with the Transaction type.