Remove model fields from Transaction and Span
Remove the "model" fields from Transaction and Span,
instead constructing the model objects in the sender.
This will make it a bit neater to support optional
distributed tracing, where we populate the model types
differently depending on whether or not distributed
tracing is enabled.

Also, allocate the transaction UUID when the transaction
is created. This is necessary for distributed tracing,
so we may as well take the hit now. We introduce a
goroutine which generates UUIDs in the background, to
minimise delays in transaction construction.