elasticapm: change TransactionOption signature
TransactionOption now accepts an unexposed
transactionOptions struct. TransactionOptions
must be defined within the elasticapm package,
so we can make assumptions about the state of
the transaction object after the options have
been applied. This will also give us a way of
specifying options like overriding the sampler,
which is not part of the transaction object.