elasticapm: drop Transaction.Ignored()
Drop the ability to ignore transactions, and
the Transaction.Ignored() method. These are
causing complications around adding support
for OpenTracing. It's also just more efficient
to keep the filtering outside of transaction
creation, so we don't have to create garbage
that is immediately thrown away. Instrumentation
modules should be responsible for that.