elasticapm: introduce ErrorID
ErrorID is a [16]byte, and is encoded as a UUID.
We may later encode it without dashes, i.e. plain
hex-encoding, depending on how the schema changes
for distributed tracing.

We now use this instead of the uuid package, and
store as [16]byte rather than as a string. The
uuid package has been removed.