module/apmot: initial tracer implementation
Implement an OpenTracing tracer on top of the
core API. Root spans create transactions, and
child spans create spans or transactions,
depending on whether or not the parent is
remote.

There is a bug in the span context which we
cannot fix until spans are top-level objects,
which will happen with the v2 API. We will
revisit this when the v2 API is supported.

Neither baggage nor logging are supported,
nor is the binary propagation format. We use
the (draft) W3C Trace-Context, for which there
is not yet a binary format.

In line with the Elastic APM Java agent's
implementation, we support several custom tags:

  - type (transaction or span type)
  - user.id, user.email, user.username
    (transaction user context)
  - result (transaction result)