module/apmgorm: introduce GORM instrumentation
Package apmgorm provides a means of instrumenting [gorm](http://gorm.io)
database operations.

By using apmgorm.Open instead of gorm.Open, we obtain a *gorm.DB which
can be used with apmgorm.WithContext. The WithContext function can be
used to propagate context (i.e. containing a transaction) to callbacks
which report spans.

We provide "dialects" packages which import the gorm/dialects namesake
packages, as well as register the apmsql drivers. The latter is required
in order to parse DSNs.