Record event type in ring buffer header
Update internal/ringbuffer, extending the header
to include a single byte tag as well as the data
size. Add an "Evicted" function field to Buffer,
which will be called with the block header for
evicted blocks.

We use this in the tracer to update the dropped
stats when the buffer is full. We also now more
accurately update the sent stats when extracting
from the buffer (writing to zlib), rather than when
writing to the buffer.