module/apmot: don't hold ended transactions/spans
Keep track of whether the underlying transaction
has been ended, and in that case use the new
Tracer.StartSpan method instead, which takes a
transaction ID rather than a transaction object.
This enables safely starting spans after the
transaction has already been ended. We no longer
refer to parent span objects when starting spans;
instead we record the span's trace context when
it is started, and use that.