module/apmredigo: introduce redigo instrumentation
We introduce module/apmredigo, which provides
top-level function for calling redis.Conn.Do
with a context, a la ctxhttp.Do. apmredigo also
provides a function (Wrap) for wrapping a redis.Conn,
diverting Do calls to the aforementioned Do-with-context
function. A Conn can be associated with a context
using the Conn.WithContext method.

Send and Receive are not instrumented. For now,
spans should be created for the higher-level
operations in which they are used (pub/sub,
transactions, etc.)