Reuse tags memory
Update model.Context.TAgs and model.SpanContext.Tags
to use the model.StringMap type, which is a slice of
key/value pairs. Storing tags is now just an append,
with no de-duplication (the server will de-duplicate
when unmarshalling). When resetting transaction and
span objects, we'll now reuse these tags slices.