Add source copyright headers
Run github.com/elastic/go-licenser to add the
copyright header to each source file. We add a
new Makefile target to check that this is kept
up to date.

Also, update LICENSE to hold the full text of
the Apache License 2.0 rather than just the
appendix, and add a NOTICE file.