module/apmot: implement Span.LogKV, Span.LogFields
Implement Span.LogKV and Span.LogFields, sending
error log events to Elastic APM, ignoring all other
logs. We only implement this for spans created
through the OT API, and not the wrapper spans
synthesised for transactions/spans created via
the native APM API.

Span.FinishWithOptions now behaves as if LogFields
were called with each of the log records in the
options. If the log record timestamp is not set,
then the span finish time will be used if specified,
and the current time otherwise.