Report error causes
We will now report complete error trees in
error.exception: the top-most exception is
the provided error, and any cause/wrapped
error will be reported in an array of causes
to the server.

The error culprit will now be calculated by
traversing the tree depth-first, returning
the first non-library frame's function name.
This will resolve issues such as #632, where
an error with a complete stack trace is
wrapped by one with only a limited stack trace.