apm: introduce ErrorDetailer interface
Introduce the ErrorDetailer interface, along with
top-level functions RegisterTypeErrorDetailer and
RegisterErrorDetailer. The former register an
ErrorDetailer with a specific error type, and the
latter registers an ErrorDetailer for all errors.

When an error is created with NewError or NewErrorLog,
we pass the error details through these ErrorDetailers,
giving instrumentation modules and application-specific
code an opportunity to extract details from custom
error types.