Introduce DetachedContext function
Introduce apm.DetachedContext, which takes a context.Context
and returns a new context.Context which is effectively a
background context, but with access to the values in the
input context.

DetachedContext can be used to maintain the trace context
required to correlate events, but where the operation is
"fire-and-forget", and should not be affected by the deadline
or cancellation of the input context.