Promote TraceContext to non-pooled types
Move TraceContext from TransactionData to
Transaction, and SpanData to Span. This
enables an ended transaction or span to
be used as the parent of a span. This is
particularly important to enable
fire-and-forget type patterns, where a
goroutine is spawned that may start a span
after the parent ends.

Span now holds a reference to its enclosing
transaction, enabling us to avoid calls to
TransactionFromContext. This may also enable
deferred transaction sampling in the future.

Transactions, spans, and errors are now all
sent to the tracer goroutine through a single
"chan tracerEvent" channel.

This change incurs a minor (~10%/~50 bytes)
increase in the number of bytes allocated
per transaction. There is no change in number
of allocs, and no significant impact on time
per operation.