module/apmot: remove workaround for ended tx
We previously worked around starting spans from
finished/ended transactions by maintaining a
reference to the span context that created the
transaction, and setting its tx property to
nil upon ending it. This was necessary because,
at the time, spans started from ended transactions
would be dropped.

We no longer drop these spans, so we can get rid
of this workaround. Now we just keep a reference
to the transaction within which a span is created,
and always use that.