module/{apmgopg,apmgoredis}: require Go 1.11
Set the minimum Go version requirement to 1.11, so
we can require the use of modules. Trying to support
older versions of Go is going to lead to a worse
experience for the majority.

Since these modules have not yet been in a release of
the agent, this is not considered a breaking change.