Introduce ELASTIC_APM_STACK_TRACE_LIMIT
We introduce a new configurable limit for the
number of stack frames to capture, defaulting
to 50.

stacktrace.AppendStacktrace has been improved
for a limit of N > 10, don't allocate space for
N callers up front; that would be wasteful if
there are mostly fewer. Instead, allocate space
for 10 and work our way up, like we do when there
is no specified limit.