Provide more control over tracer creation
Introduce NewTracerOptions and TracerOptions, providing
a means of specifying the initial tracer configuration
(currently just a limited subset).

We now use NewTracerOptions to start a Tracer with an
initial non-default Transport in all of the tests,
rather than starting with the default and switching
after the fact.

Importing apmtest will now update apm.DefaultTracer to
DiscardTracer, calling the initial value's Close method
first.

We also introduce apmtest.NewRecordingTracer and
deprecate direct use of transporttest.RecorderTransport
in tests. This is in order to encourage tests to use
apmtest, updating the default tracer as described above.