Add SetLabel, deprecate SetTag
Add Context.SetLabel and SpanContext.SetLabel,
which accept interface{} values. The value may
be a nil, string, bool, or numerical type. Any
other type will be sent as a string using
fmt.Sprint.

The SetTag methods remain, but marked as being
deprecated, with a notice that they will be
removed in a future major version.