Add "tracestate" propagation
Add the field TraceContext.State, which has the new
type TraceState. TraceState is an opaque type, which
internally contains a list of TraceStateEntry objects.

When starting a transaction with StartTransactionOptions,
the TraceContext.State field will only be considered if
both TraceContext.Trace and TraceContext.State are valid.
If this is true, then the state will be stored on the
transaction, and spans created using the transaction.

Server instrumentation in apmhttp, apmgrpc, and apmot
will extract and parse "tracestate" headers, while client
instrumentation will inject "tracestate" headers.

Add missing docs for `ELASTIC_APM_USE_ELASTIC_TRACEPARENT_HEADER`.