Add support for errors returning *runtime.Frames
The standard library has a built-in concept of stack traces in the
runtime package. While there is no standard of how errors are tied to
stack traces, the current implementation of requires either stack traces
from this library directly, or from github.com/pkg/errors.

With this change, any library that implements the following interface
would have their stack traces supported within APM:

    type runtimeStackTracer interface {
      StackTrace() *runtime.Frames
    }

The advantage to this is that now an error-handling library with
stack-trace support can make itself compatible with this library without
requiring go.elastic.co/apm or github.com/pkg/errors as a dependency.
They only have to import the "runtime" package.

For reference, here is one such library:
<github.com/morningconsult/serrors>