module/apmot: fix nil-pointer deref in Inject
If the SpanContext passed to Inject does not
contain a Transaction, then Inject would panic
due to a nil pointer dereference. We were using
the transaction to check if we should inject
the legacy header. Since this is not a dynamic
configuration, we can just use the tracer instead,
by adding a Tracer.ShouldPropagateLegacyHeader
method.