Record {transaction,sample}.sample_rate (#804)
* Record {transaction,sample}.sample_rate

Introduce the new ExtendedSampler interface, which
Samplers may implement to return the effective
sampling rate. This is implemented by the built-in
ratioSampler.

When starting a root transaction we now call the
ExtendedSampler method if implemented, and store
the effective sampling rate in the transaction's
tracestate under our "es" vendor key.

When receiving tracestate, we parse our "es" vendor
value and extract the sample rate.

When encoding transactions and spans we record the
sample rate (from tracestate) in the transaction and
span events.