Update sample rate precision handling
We already had sampling rate rounding,
but it was using the older proposal of
3 decimal places, and did not clamp the
lower bound.

Updated to use the same logic in both the
ratio sampler, and the implementation-
independent interpretation of the sample
rate.

Also, use the "round" function instead
of math.Round directly, to support older
versions of Go.