scripts/jenkins: fix CI builds on 1.16
In 1.16, by default go will complain if go.mod
is missing dependencies. In CI this is occurring
due to the installation of tools that are not
present in the top-level go.mod.

To avoid adding dependencies for agent consumers,
introduce a go.mod specifically for installing
tools used only in CI.