handle path/hosted bucket style names
There are two names the bucket name can get
encoded into the request URL, either as a
subdomain prepended to the host, or as the first
member in the path. Handle both of these, as we
don't have control over which will be used.

Even if explicitly set in the config to use the
hosted bucket style, it might only be able to do
the path style.

https://github.com/aws/aws-sdk-go/blob/main/aws/config.go#L118-L127