BodyCapturer: Fix NPE on benchmark when nil Context.request
Fixes a nil pointer exception on the benchmark `BenchmarkBodyCapturer`.
The `Context.request` field wasn't set in the benchmark and is being
passed in the `BodyCapturer.setContext()` call as nil, causing an NPE
when `req.PostForm` is accessed.

The bug seems to have been introduced in #906.

This might not be the right fix since I'm not very familiar with the
codebase.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>