Address review comments
Adds a new `parentID` field to both `Transaction` and `Span` structures.
The `parentID` is only populated at start time and remains immutable
throughout the lifecycle of the Transaction or Span, making it possible
to retrieve the `parentID` for either after they have ended.
However, the `ParentID()` call might return a zero value `SpanID` when
a trace has no parent.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>