Remove `spanBuffer`, simplify approach
Removes the `spanBuffer`, adding a new field to `Transaction` and `Span`
instead and handling the caching directly on that field. `evictCache` is
used to flush the `TransactionData.cache` and `SpanData.cache`

Also simplifies a few functions and uses a `strings.Builder` with the
total capacity set beforehand to update the `span.Name` since that seems
to be the fastest and most efficient way to concatenate the two strings,
the allocation is unavoidable.

Signed-off-by: Marc Lopez Rubio <marc5.12@outlook.com>