Address review comments
Adds a new `s.compressOrEvictCache`, and simplifies `s.attemptCompress`.
The new method compresses or evicts the span into / from the cache.

Also modifies the `s.evict()` to set the cached span compression options
to `false` so the span doesn't go through the compression flow again and
causes infinite loop.

Last, it splits the compression and enqueuing portion of `s.End()` into
a new `s.end()` method which is called only for spans that have been
evicted from the cache.

Signed-off-by: Marc Lopez Rubio <marc5.12@outlook.com>