span (experimental): Compress short exit spans (#1134)
Implements the span compression algorithm described in elastic/apm#432.
The implementation pretty close to what the spec describes, with a few
modifications due to the differences in the Go agent implementation. The
span compression is experimental and is disabled by default with the
default thresholds for the different compression strategies set:

- `ELASTIC_APM_SPAN_COMPRESSION_ENABLED=false`
- `ELASTIC_APM_SPAN_COMPRESSION_EXACT_MATCH_MAX_DURATION=50ms`
- `ELASTIC_APM_SPAN_COMPRESSION_SAME_KIND_MAX_DURATION=5ms`

The implementation uses a new field (cache `*Span`) in `TransactionData`
and `SpanData` which is used to cache compressed or compression eligible
spans in that field. An additional `composite` field has also been added
to the `SpanData` and `TransactionData`.

The algorithm states that any exit span that is lower or equal duration
than the set `ELASTIC_APM_SPAN_COMPRESSION_EXACT_MATCH_MAX_DURATION` or
`ELASTIC_APM_SPAN_COMPRESSION_SAME_KIND_MAX_DURATION` with a destination
service set is compressable into a composite span using the appropriate
strategy. Compressable spans need to be "exit" and siblings which **have
not** propagated their context downstream. Sibling spans are spans that
share the same parent span (or transaction).

Spans can be compressed with `same_kind` or `exact_match` as a strategy,
and their sum is the aggregated duration of of all
When a compressed span has been compressed into a composite using the
`same_kind` strategy, its name is mutated to `Calls to <span.Type>`.

Spans will be compressed into a composite only when the siblings are
consecutive and any compression attempt that doesn't meet that,
will cause the cache to be evicted.

Additionally, two helper functions have been added under the `apmtest`:
package: `WriteTraceTable` and `WriteTraceWaterfall`, which help with
understanding why a test is failing if they are and debugging those.

Signed-off-by: Marc Lopez Rubio <marc5.12@outlook.com>