span: Add config setting `exit_span_min_duration`
Adds a new configuration setting which drops discardable exit spans when
their duration is lower or equal than `exit_span_min_duration`. If span
compression is enabled, the duration of the composite is considered.

This setting drops the spans after they have ended in contrast with the
existing `transaction_max_spans` which drops them from the start.

The existing a new `configutil.ParseDurationOptions` has been introduced
which allows specifying a custom unit as the minimum duration.

Signed-off-by: Marc Lopez Rubio <marc5.12@outlook.com>