Drop discardable spans after they end
Updates the `end()` private method and moves the compression attempts to
the Public function so that every iteration of `cachedSpan.end()` does
not have to go through that logic again.

Moves the `s.dropped` logic to the private `end` function since it now
needs to check if the span is discardable on every span end, but only
after the compression cache has been evicted.

Most of the lock acquisition has been moved out of the `end()` function
and onto the public `End()`. The main reason is because the transaction
may call `cachedSpan.End()` with a different set of locks acquired vs
`span.End()`.

Last, fixes a small bug where `transaction.dropped_spans_stats.outcome`
was empty if the outcome hadn't been set in the modelwriter.

Signed-off-by: Marc Lopez Rubio <marc5.12@outlook.com>