tracer: Drop unsampled txs when APM Server >= 8.0
Extends the HTTP Transport with a new method that queries the remote APM
Server `/` endpoint to obtain its version. This allows the tracer to
drop unsampled transactions when the APM Server to which it sends traces
is version `8.0.0` or higher.

The version will be cached in a local variable until the APM Server URL
changes which will refresh the APM Server's version.

Signed-off-by: Marc Lopez Rubio <marc5.12@outlook.com>