tracer: Drop unsampled txs when APM Server >= 8.0 (#1208)
Extends the HTTP Transport with a new method that queries the remote APM
Server `/` endpoint to obtain its version. Allowing the tracer to drop
unsampled transactions when the APM Server is version `8.0.0` or higher.

The version will be cached in a local variable until the APM Server URL
changes which will refresh the APM Server's version. The tracer launches
a goroutine to update the version when it is instantiated and refreshes
the remote server version every 10 seconds.

Any SendStream errors will cause the version cache to be invalidated and
will be refreshed the next time the ticker runs. It may cause unsampled
to be sent to the APM Server while the cache is invalid.

Signed-off-by: Marc Lopez Rubio <marc5.12@outlook.com>