tracer: Parse global labels per tracer
This patch removes the `globalLabel` variable and one-time parsing of
the global labels, in favor of parsing the global labels individually
on tracer construction. This has the advantage that tracers that are
created after start-up will parse environment variables at that time
and not only when the `apm` package is imported.

Encountered while writing a systemtest.

Signed-off-by: Marc Lopez Rubio <marc5.12@outlook.com>