fix: mark spans representing calls to external services as exit spans
Exit spans are spans that describe a call to an external service,
such as an outgoing HTTP request or a call to a database.

Marking a span as an exit span also triggers service target fields
and destination fields inference.
Update tests to reflect the new fields.

Disable compression in some tests as exit spans can be aggregated
based on certain fields, leading to unexpected results.

Additionally, make sure exit spans are not being used as parents of
other spans as that would cause the spans to be dropped.