fix: allow exit spans to have child spans with the same type and subtype
Do not silently drop child spans of exit spans. The spec allows for exit
spans to have child spans if they have the same type and subtype.
One example of this is http trace requests being child spans of an exit
span.

Update exit span test to make sure the feature is working as intended.