fix: allow exit spans to have child spans with the same type and subtype (#1320)
* fix: allow exit spans to have child spans with the same type and subtype

Do not silently drop child spans of exit spans. The spec allows for exit
spans to have child spans if they have the same type and subtype.
One example of this is http trace requests being child spans of an exit
span.

Update exit span test to make sure the feature is working as intended.

* fix: update code to compile on older versions of go

* fix: move exit span parent check in span.end

Type/SubType is mutable. Move validation inside span.end and drop span
according to the spec.

* fix: remove destination context from child spans of exit spans

According to the spec:

These spans MUST NOT have any destination context, so that there's no effect on destination metrics.

Update the code accordingly to remove destination context from them.

* test: update exit span test for code changes

Moving the check to span.end introduced a small change in behaviour:
child spans of exit spans are not dropped immediately. Update test case
to account for that

* fix: update the correct field for model destination context

Co-authored-by: Andrew Wilkins <axwalk@gmail.com>

* fix: follow the proper flow for span ending/dropping

Do not null out the tracer, instead call the method and follow
the proper flow.
The destination context is set to null to prevent the spans from
being aggregated in the dropped spans metrics.

Co-authored-by: Andrew Wilkins <axwalk@gmail.com>