Fix nil check during unwrapping.
Instead of aborting after encountering a `nil` error and thus missing other potential causes, add missing nil check in
the `Unwrap` method that takes a slice.

Also include a test for actually exercising this code and reproducing the problem seen in the wild (in Buildkite).
Thanks to @kruskall for providing the test and the excellent suggestion.